refactor: ownerUserId 필드를 Integer에서 Long으로 변경

+++ src/main/java/co/kr/muldum/domain/file/model/File.java
@@ -35,7 +35,7 @@ public class File {
   private UserType ownerUserType;
 
   @Column(name = "owner_user_id", nullable = false)
-  private Integer ownerUserId;
+  private Long ownerUserId;
 
   @UpdateTimestamp
   @Column(name = "updated_at", nullable = false)
